test(roborazzi): golden the editor with the formatting row gone (#331)
The setting that removes the standing `Aa` row shipped with one instrumented
assertion — the node does not exist — and a hand check on an emulator. Neither
can see the thing the setting exists to fix. The complaint in #328 was not the
button, it was the 48dp strip of empty screen it holds above the keyboard, so
what "off" has to produce is an absence with nothing left behind. A layout
change that left the row's height while dropping its content would keep the
node absent and pass, and `EditorFormattingControlsSnapshotTest` cannot cover
it either: in this state the row is not mounted, so there is nothing there to
photograph.

That makes it a screen-level golden, and the screen builds its own
`AppSettingsRepository`, which a Robolectric test cannot put into a chosen
state — the `preferencesDataStore` delegate caches one instance per JVM and
DataStore's `File.renameTo` writes fail on Windows after the first, which is
why the repository already has an internal DataStore constructor (#158).
`EditorScreen` now takes the repository as a defaulted parameter, so the same
seam that makes the repository testable makes the screen it feeds testable too.
Nothing else changes: the default is the process-wide store, and the existing
golden still composes the screen exactly as the app builds it.

`InMemoryPreferencesDataStore` moves out of `AppSettingsRepositoryLockTest`,
where it was a private copy also duplicated in `AppSettingsRepositoryListTest`,
into one class the snapshot test can use as well.

The new test asserts the premise before capturing: an image cannot say why it
looks the way it does, and this one is defined by an absence, so a capture that
happened before the setting reached the screen would look plausible and pin the
wrong state. Both tests check the formatting node — present by default, absent
with the setting off — and only then photograph the surface.

Golden recorded on the Linux runner via the `record_roborazzi` dispatch, since
Windows font hinting produces images CI then rejects.
