+++ app/src/test/java/com/markleaf/notes/data/settings/AppSettingsRepositoryListTest.kt
@@ -1,14 +1,7 @@
 package com.markleaf.notes.data.settings
 
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.emptyPreferences
 import com.markleaf.notes.core.text.NoteTitleSource
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.sync.Mutex
-import kotlinx.coroutines.sync.withLock
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertEquals
 import org.junit.Test
@@ -19,20 +12,6 @@ import org.junit.Test
  * the store is injected rather than the production singleton.
  */
 class AppSettingsRepositoryListTest {
-
-    private class InMemoryPreferencesDataStore : DataStore<Preferences> {
-        private val state = MutableStateFlow(emptyPreferences())
-        private val mutex = Mutex()
-        override val data: Flow<Preferences> = state
-
-        override suspend fun updateData(
-            transform: suspend (t: Preferences) -> Preferences
-        ): Preferences = mutex.withLock {
-            val updated = transform(state.value)
-            state.value = updated
-            updated
-        }
-    }
 
     private val store = InMemoryPreferencesDataStore()
     private val repo = AppSettingsRepository(store)
